# api/v1alpha1/resource_manager/types.gen.go
// Package resource_manager provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.1-0.20260318123712-00a90b7a03f4 DO NOT EDIT.
package resource_manager

import (
	"time"
)

// Defines values for ServiceTypeInstanceDeletionStatus.
const (
	FAILED          ServiceTypeInstanceDeletionStatus = "FAILED"
	PENDINGPROVIDER ServiceTypeInstanceDeletionStatus = "PENDING_PROVIDER"
	SCHEDULED       ServiceTypeInstanceDeletionStatus = "SCHEDULED"
)

// Valid indicates whether the value is a known member of the ServiceTypeInstanceDeletionStatus enum.
func (e ServiceTypeInstanceDeletionStatus) Valid() bool {
	switch e {
	case FAILED:
		return true
	case PENDINGPROVIDER:
		return true
	case SCHEDULED:
		return true
	default:
		return false
	}
}

// Error RFC 7807 compliant error response
type Error struct {
	// Detail Human-readable explanation specific to this occurrence
	Detail *string `json:"detail,omitempty"`

	// Instance URI reference for this specific error occurrence
	Instance *string `json:"instance,omitempty"`

	// Status HTTP status code
	Status *int `json:"status,omitempty"`

	// Title Short human-readable summary of the problem
	Title string `json:"title"`

	// Type URI reference identifying the error type
	Type string `json:"type"`
}

// Health Health status singleton resource
type Health struct {
	// Path Canonical path of the resource
	Path *string `json:"path,omitempty"`

	// Status Health status
	Status *string `json:"status,omitempty"`
}

// ServiceTypeInstance Full service type instance resource representation
type ServiceTypeInstance struct {
	// CreateTime Timestamp when the instance was first created
	CreateTime *time.Time `json:"create_time,omitempty"`

	// DeletionStatus Deletion status for deferred deletions. Absent for active
	// instances. SCHEDULED indicates the instance is queued for cleanup.
	// FAILED indicates the cleanup has exceeded maximum retries.
	// PENDING_PROVIDER indicates the instance is waiting for its
	// provider to become healthy before cleanup is retried.
	DeletionStatus *ServiceTypeInstanceDeletionStatus `json:"deletion_status,omitempty"`

	// Id Unique identifier for the Service Type Instance
	Id *string `json:"id,omitempty"`

	// Path Resource path identifier
	Path *string `json:"path,omitempty"`

	// ProviderName Name of the provider
	ProviderName string `json:"provider_name"`

	// Spec Service specification following one of the supported service type
	// schemas (VMSpec, ContainerSpec, DatabaseSpec, or ClusterSpec).
	Spec map[string]interface{} `json:"spec"`

	// Status Status of the instance
	Status *string `json:"status,omitempty"`

	// UpdateTime Timestamp when the instance was last updated
	UpdateTime *time.Time `json:"update_time,omitempty"`
}

// ServiceTypeInstanceDeletionStatus Deletion status for deferred deletions. Absent for active
// instances. SCHEDULED indicates the instance is queued for cleanup.
// FAILED indicates the cleanup has exceeded maximum retries.
// PENDING_PROVIDER indicates the instance is waiting for its
// provider to become healthy before cleanup is retried.
type ServiceTypeInstanceDeletionStatus string

// ServiceTypeInstanceList Paginated list of instances
type ServiceTypeInstanceList struct {
	Instances *[]ServiceTypeInstance `json:"instances,omitempty"`

	// NextPageToken Token for retrieving the next page of results
	NextPageToken *string `json:"next_page_token,omitempty"`
}

// InstanceIdPath defines model for InstanceIdPath.
type InstanceIdPath = string

// ListInstancesParams defines parameters for ListInstances.
type ListInstancesParams struct {
	// Provider Filter service provider
	Provider *string `form:"provider,omitempty" json:"provider,omitempty"`

	// ServiceType Filter instances by service type
	ServiceType *string `form:"service_type,omitempty" json:"service_type,omitempty"`

	// ShowDeleted If true, soft-deleted instances are included in the results
	// alongside active instances. Defaults to false.
	ShowDeleted *bool `form:"show_deleted,omitempty" json:"show_deleted,omitempty"`

	// MaxPageSize Maximum number of results per page
	MaxPageSize *int `form:"max_page_size,omitempty" json:"max_page_size,omitempty"`

	// PageToken Token for pagination
	PageToken *string `form:"page_token,omitempty" json:"page_token,omitempty"`
}

// CreateInstanceParams defines parameters for CreateInstance.
type CreateInstanceParams struct {
	// Id Optional ID for service type instance
	Id *string `form:"id,omitempty" json:"id,omitempty"`
}

// DeleteInstanceParams defines parameters for DeleteInstance.
type DeleteInstanceParams struct {
	// Deferred When true, deletion failures are deferred to a cleanup queue
	// instead of returning an error. Used during rehydration flows.
	Deferred *bool `form:"deferred,omitempty" json:"deferred,omitempty"`
}

// GetInstanceParams defines parameters for GetInstance.
type GetInstanceParams struct {
	// ShowDeleted If true, soft-deleted instances are returned instead of 404.
	// Defaults to false.
	ShowDeleted *bool `form:"show_deleted,omitempty" json:"show_deleted,omitempty"`
}

// CreateInstanceJSONRequestBody defines body for CreateInstance for application/json ContentType.
type CreateInstanceJSONRequestBody = ServiceTypeInstance